+++ tests/unit/models/test_auth.py
@@ -38,7 +38,19 @@ class TestAuth(UnitTest):
     def test_url__installed_app(self):
         url = installed_app().auth.url(['dummy scope'], 'dummy state')
         assert 'client_id=dummy+client' in url
+        assert 'duration=permanent' in url
         assert 'redirect_uri=https%3A%2F%2Fdummy.tld%2F' in url
+        assert 'response_type=code' in url
+        assert 'scope=dummy+scope' in url
+        assert 'state=dummy+state' in url
+
+    def test_url__installed_app__implicit(self):
+        url = installed_app().auth.url(['dummy scope'], 'dummy state',
+                                       implicit=True)
+        assert 'client_id=dummy+client' in url
+        assert 'duration=temporary' in url
+        assert 'redirect_uri=https%3A%2F%2Fdummy.tld%2F' in url
+        assert 'response_type=token' in url
         assert 'scope=dummy+scope' in url
         assert 'state=dummy+state' in url
 
@@ -47,8 +59,13 @@ class TestAuth(UnitTest):
         assert 'client_id=dummy+client' in url
         assert 'secret' not in url
         assert 'redirect_uri=https%3A%2F%2Fdummy.tld%2F' in url
+        assert 'response_type=code' in url
         assert 'scope=dummy+scope' in url
         assert 'state=dummy+state' in url
+
+    def test_url__web_app__implicit(self):
+        with pytest.raises(ClientException):
+            web_app().auth.url(['dummy scope'], 'dummy state', implicit=True)
 
     def test_url__web_app_without_redirect_uri(self):
         reddit = Reddit(client_id='dummy client', client_secret='dummy secret',